% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpNE.R
\name{DAISIE_DE_trait_logpNE}
\alias{DAISIE_DE_trait_logpNE}
\title{This function calculates the likelihood of observing a non-endemic lineage with specified species trait states,}
\usage{
DAISIE_DE_trait_logpNE(
  datalist,
  brts,
  parameter,
  trait,
  num_observed_states,
  num_hidden_states,
  trait_mainland_ancestor = NA,
  status,
  sampling_fraction,
  Mainland_pool_size_vec = NULL,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "ode45",
  rcpp_methode = "odeint::runge_kutta_cash_karp54",
  use_Rcpp = 2
)
}
\arguments{
\item{datalist}{a list containing in the first entry: island age and the
number of species not present, then from entry 2 until n it contains
individual datasets for which to estimate the loglikelihood together, each
dataset has the following fields: 1) colonist_name, 2) b ranching_times,
3) stac, 4) missing_species, 5) type1or2, 6) traits, 7) sampling fraction,
8) phylogeny}

\item{brts}{branching times, being: island age, colonization time, branching
times}

\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{trait}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
`vignette("starting_secsse", package = "secsse")`. When providing a
`multiPhylo` set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{num_observed_states}{number of observed states}

\item{num_hidden_states}{number of hidden states}

\item{trait_mainland_ancestor}{if mainland is TRUE, sets the trait of the
mainland ancestor}

\item{status}{indicator for type of inference required}

\item{sampling_fraction}{vector that states the sampling proportion per
trait state. It must have as many elements as there are trait states. When
using a `multiPhylo` object, sampling fraction should be list where each
entry in the list corresponds to the sampling proportion for each tree.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{used integration method, default is ode45}

\item{rcpp_methode}{Used integration methode when using Rcpp to integrate,
options available are:
 `"odeint::runge_kutta_cash_karp54"`, `"odeint::runge_kutta_fehlberg78"`,
 `"odeint::runge_kutta_dopri5"`, `"odeint::bulirsch_stoer"` and
 `"odeint::runge_kutta4"`. Default method is:
 `"odeint::runge_kutta_cash_karp54"`.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree
likelihood, value 2 = use Rcpp for the tree likelihood AND the ODEs along
the branches}
}
\description{
This function calculates the likelihood of observing a non-endemic lineage with specified species trait states,
}
\examples{
library(DAISIE)
data("Galapagos_datalist")
datalist <- Galapagos_datalist
datalist[[1]]$M0 <- 500
datalist[[1]]$M1 <- 400
i <- 3
brts <- datalist[[i]]$branching_times
trait <- 0

parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    .001,    0.005,  0,
    .001,    0,    0.000,0.005,
    0.005,    000,    0,  0.005,
    0,   0.005,  0.005,0.00
  ), nrow = 4, byrow = TRUE),
  1
)
brts <- datalist[[i]]$branching_times
DAISIE_DE_trait_logpNE(
  datalist                = datalist,
  brts                    = brts,
  trait                   = trait,
  status                  = 4,
  parameter               = parameter,
  sampling_fraction       = c(1,1),
  trait_mainland_ancestor = c(1,0),
  num_observed_states     = 2,
  num_hidden_states       = 2,
  atol                    = 1e-15,
  rtol                    = 1e-15,
  methode                 = "ode45",
  use_Rcpp                = 2)


}
