% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpNE_max_min_age_hidden.R
\name{DAISIE_DE_trait_logpNE_max_min_age_hidden}
\alias{DAISIE_DE_trait_logpNE_max_min_age_hidden}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logpNE_max_min_age_hidden(
  brts,
  parameter,
  trait,
  num_observed_states,
  num_hidden_states,
  trait_mainland_ancestor = NA,
  status,
  atol = 1e-15,
  rtol = 1e-15,
  sampling_fraction = c(1, 1),
  methode = "ode45",
  rcpp_methode = "odeint::runge_kutta_cash_karp54",
  use_Rcpp = 2
)
}
\arguments{
\item{brts}{branching times, being: island age, colonization time, branching
times}

\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{trait}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
`vignette("starting_secsse", package = "secsse")`. When providing a
`multiPhylo` set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{num_observed_states}{number of observed states}

\item{num_hidden_states}{number of hidden states}

\item{trait_mainland_ancestor}{if mainland is TRUE, sets the trait of the
mainland ancestor}

\item{status}{indicator for type of inference required}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{sampling_fraction}{vector that states the sampling proportion per
trait state. It must have as many elements as there are trait states. When
using a `multiPhylo` object, sampling fraction should be list where each
entry in the list corresponds to the sampling proportion for each tree.}

\item{methode}{used integration method, default is ode45}

\item{rcpp_methode}{Used integration methode when using Rcpp to integrate,
options available are:
 `"odeint::runge_kutta_cash_karp54"`, `"odeint::runge_kutta_fehlberg78"`,
 `"odeint::runge_kutta_dopri5"`, `"odeint::bulirsch_stoer"` and
 `"odeint::runge_kutta4"`. Default method is:
 `"odeint::runge_kutta_cash_karp54"`.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree
likelihood, value 2 = use Rcpp for the tree likelihood AND the ODEs along
the branches}
}
\description{
this function calculates the likelihood of observing a singleton species on
an island with the trait state `i`, either non-endemic or rendered endemic
by a trait change, and for which only the estimated maximum age of
colonization is known.
}
\examples{
# load DAISIE package and data
library(DAISIE)

parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    .001,    0.005,  0,
    .001,    0,    0.000,0.005,
    0.005,    000,    0,  0.005,
    0,   0.005,  0.005,0.00
  ), nrow = 4),
  1
)
status <- 8


DAISIE_DE_trait_logpNE_max_min_age_hidden(
  brts                  = c(4, 3.999, 0.0001),
  trait                 = 0,
  status                = 8,
  parameter             = parameter,
  num_observed_states   = 2,
  num_hidden_states     = 2,
  trait_mainland_ancestor = c(1,0),
  sampling_fraction       = c(1,1),
  atol                  = 1e-15,
  rtol                  = 1e-15,
  methode               = "ode45",
  rcpp_methode = "odeint::runge_kutta_cash_karp54"
)
}
